+++ MODELS/automobile.h
@@ -12,7 +12,7 @@ public:
      * @param for MotoreCombustione(capacitaSerbatoio, litriSerbatoio, tipoCarburante, cilindrata, emissioni)
      */
     Automobile(string targa, string posizione, double km, double capacitaSerbatoio, double litriSerbatoio,
-               Carburante tipoCarburante, u_int cilindrata, u_int emissioni, u_int numeroPosti = 2, u_int ingombro = 3);
+               Carburante tipoCarburante, unsigned short cilindrata, unsigned short emissioni, unsigned short numeroPosti = 2, unsigned short ingombro = 3);
 
 };
