+++ MODELS/automobileelettrica.h
@@ -11,9 +11,9 @@ public:
      * @param for MotoreElettrico(potenza, numeroMotori, inCarica, capacitaBatteria, capacitaAttuale, caricaSupportata, colonninaAttuale)
      */
     AutomobileElettrica(string targa, string posizione, double km,
-                        u_int potenza, u_int numeroMotori, double capacitaBatteria, double capacitaAttuale, VelocitaRicarica caricaSupportata,
-                        bool inCarica = false, Colonnina colonninaAttuale = nessuna,
-                        u_int numeroPosti = 2, u_int ingombro = 3);
+                        unsigned short potenza, unsigned short numeroMotori, double capacitaBatteria, double capacitaAttuale,
+                        VelocitaRicarica caricaSupportata, bool inCarica = false, Colonnina colonninaAttuale = nessuna,
+                        unsigned short numeroPosti = 2, unsigned short ingombro = 3);
 
 };
