+++ MODELS/monopattino.cpp
@@ -2,7 +2,7 @@
 
 Monopattino::Monopattino(string targa, string posizione, double km, double capacitaBatteria, double capacitaAttuale,
                          VelocitaRicarica caricaSupportata, unsigned short potenza, unsigned short numeroMotori, bool inCarica, Colonnina colonninaAttuale,
-                         u_int numeroPosti, u_int ingombro):
+                         unsigned short numeroPosti, unsigned short ingombro):
     Veicolo(targa, posizione, km, numeroPosti, ingombro),
     MotoreElettrico(targa, posizione, km, numeroPosti, ingombro,  potenza,  numeroMotori, inCarica, capacitaBatteria, capacitaAttuale, caricaSupportata, colonninaAttuale)
 {}
