+++ MODELS/motoreelettrico.cpp
@@ -1,9 +1,11 @@
 #include "MODELS/motoreelettrico.h"
 
-MotoreElettrico::MotoreElettrico(string targa, string posizione, double km, u_int numeroPosti, u_int ingombro,
-                                 u_int potenza, u_int numeroMotori, bool inCarica, double capacitaBatteria, double capacitaAttuale, VelocitaRicarica caricaSupportata, Colonnina colonninaAttuale):
+MotoreElettrico::MotoreElettrico(string targa, string posizione, double km, unsigned short numeroPosti, unsigned short ingombro,
+                                 unsigned short potenza, unsigned short numeroMotori, bool inCarica, double capacitaBatteria,
+                                 double capacitaAttuale, VelocitaRicarica caricaSupportata, Colonnina colonninaAttuale):
     Veicolo(targa, posizione, km, numeroPosti, ingombro),
-    _potenza(potenza), _numeroMotori(numeroMotori) , _inCarica(inCarica), _capacitaBatteria(capacitaBatteria), _capacitaAttuale(capacitaAttuale), _caricaSupportata(caricaSupportata), _colonninaAttuale(colonninaAttuale) {}
+    _potenza(potenza), _numeroMotori(numeroMotori) , _inCarica(inCarica), _capacitaBatteria(capacitaBatteria), _capacitaAttuale(capacitaAttuale),
+    _caricaSupportata(caricaSupportata), _colonninaAttuale(colonninaAttuale) {}
 
 MotoreElettrico::~MotoreElettrico() {}
